+++ NEWS.md
@@ -60,6 +60,10 @@ Standard library changes
 
 #### Profile
 
+#### Random
+
+* `shuffle` now may take an argument of `NTuple` value ([#56906]).
+
 #### REPL
 
 #### Test